Ajustes finais

+++ src/main/java/br/upf/ads/topicos/converter/TemplateConverter.java
@@ -17,7 +17,7 @@ public class TemplateConverter implements Converter{
 		if (value != null && value.trim().length() > 0) {
 			try {
 				EntityManager em = JpaUtil.getInstance().getEntityManager();
-				Template ret = em.find(Template.class, Long.parseLong(value));
+				Template ret = em.find(Template.class, Integer.parseInt(value));
 				em.close();
 				return ret;
 			} catch (NumberFormatException e) {
